Update auth.php to lowercase username
Lowercase username when checking if user already exists as create_user_record lowercases it before insertion.

Postgresql default install for Moodle uses case sensitive collation.

If auto-create accounts is enabled, the account will be created and login will work for that session. As soon as you hit logout and try and login again, you receive a DML exception as Moodle attempts to insert the record to the DB with 'create_user_record' - it violates the unique index, as it already exists in the lowercase form.

Code is cribbed from auth/manual, but we don't check the mnet id, may be bug or feature.